refactor: move boundary-spacing.ts into src/browser/dom/
Step 1 of splitting src/browser/pangu.ts so src/browser root holds
only pangu.ts and pangu.umd.ts, per tmp/split-pangu-ts-handoff.md.
Updates the import in pangu.ts and the test import paths accordingly.
